Add files via upload
Hello to you, Philip. I have just uploaded and pushed (to the GitHub OldWeather3 master repository) the final, quality-controlled IMMA data files for the OldWeather3 Albatross (1884) ship and the OldWeather3 Thetis ship. Both of these two IMMA data files are now ready for you to accept and add into the master repository if they are not located there already and if this is a step that you need to take to finish the process. Thank you.  -Larry Spencer