Add files via upload
Dear Philip,

Hello and pleasant day to you. I have the final version of the IMMA data file for the OldWeather3 Jamestown (1886) ship ready to submit to you so that you could accept and transfer it into the OldWeather3 master repository here on the GitHub website. I have uploaded this particular data file to this GitHub website for you so that you could have access to it and put it into the OldWeather3 master repository. Please let me know if you have any problems at all with accessing this data file. The full name of it is Jamestown_1886.imma  Thank you and have a great day.

Kindest Regards,
Larry Spencer