Add files via upload
Dear Philip,

Hello and very pleasant day to you, Philip. I want to kindly inform you that I have just uploaded the final IMMA data file for the OldWeather3 Jamestown (1866) ship to the GitHub website, and I have submitted a "Pull Request" to you so that it can be accepted and merged into the OldWeather3 Master Repository here on GitHub. I completely realize and am aware that there is already a final IMMA data file for the OldWeather3 Jamestown (1866) ship, which is named Jamestown_1866.imma, located in the OldWeather3 Master Repository and that it had been merged into that particular location approximately one year ago. However, earlier today I was carefully reviewing the most recent dates of modification of all of the final IMMA data files that I have created for OldWeather3 ships so far throughout the past few years, and I noticed that the most recent date of modification of the final IMMA data file that I had created for the OldWeather3 Jamestown (1866) ship was earlier this year back in January of 2018, which is more recent than the time (approximately one year ago) when the previous Jamestown_1866.imma file was accepted and merged into the OldWeather3 Master Repository. Therefore, to simply account for this situation and to serve as a precautionary measure, I have uploaded and submitted this particular data file to you again here on GitHub so that this up-to-date Jamestown_1866.imma file is included in the OldWeather3 Master Repository. Would it be any problem at all to have two IMMA data files, which are named Jamestown_1866.imma, with the same name to exist and be located in the OldWeather3 Master Repository here on GitHub? Would there be any way to replace the already existing Jamestown_1866.imma file in the OldWeather3 Master Repository with the up-to-date Jamestown_1866.imma file that I have uploaded and submitted to you in this "Pull Request" here on GitHub today? Please take care, and I hope that you have a fantastic day. Thank you.

Kindest Regards,
Larry Spencer