Add files via upload
Dear Philip - Hello and very good Friday morning to you, Philip. I wanted to kindly inform you that I needed to go back and make a slight/minor modification inside of the Albatross_1890.imma file that I had uploaded here on GitHub a couple of days ago. Could I please make a small request to you? If it is possible for you to do so, could you please accept and merge this newly-uploaded Albatross_1890.imma file into the OldWeather3 Master Repository here on GitHub and replace the other IMMA file with this new IMMA file for the Albatross (1890) ship? This new IMMA file is only slightly different from the other IMMA file, and the vast majority of it is exactly the same as the other one because the modification that I needed to make inside of the file was quite minor. I do apologize to you in advance for any inconvenience that this might cause, and thank you very much in advance for kindly taking care of this for me (if it is possible to do so). I hope that you have a fantastic day. Thank you, again.

Kindest Regards,
Larry Spencer