Add files via upload
Friday, September 13th, 2019 - Commit Message Written by Larry Spencer

This is a commit message in regards to the IMMA data file for the OldWeather3 Albatross (1890) ship. A final IMMA data file has been created and is now available for the OldWeather3 Albatross (1890) ship. I have specifically indicated that this is a final IMMA data file for this ship by including the word "Final" in the name of the file. The name of the file is Albatross_1890_Final.imma, and I have uploaded this particular Albatross_1890_Final.imma file to GitHub. This final IMMA data file has been uploaded to GitHub, and it is now available and ready to be accepted/added into the OldWeather3 Master Repository from my LawrenceSpencer Master Repository on GitHub. Just to reiterate, this file is the final IMMA data file for the OldWeather3 Albatross (1890) ship.

Kindest Regards,
Larry Spencer