Add files via upload
Friday, September 13th, 2019 - Commit Message Written by Larry Spencer

This is a commit message in regards to the IMMA data file for the OldWeather3 Jamestown (1844) ship. A preliminary IMMA data file has been created and is now available for the OldWeather3 Jamestown (1844) ship. This IMMA data file is only a preliminary/non-final version of the file because it contains the quality-controlled geographical positions and meteorological observations data for only the time period of June 1861 through September 1865 of the OldWeather3 Jamestown (1844) ship's completed voyage. Since the file contains the quality-controlled geographical positions/meteorological observations data for only the last 4 years and 3 months (June 1861 - September 1865) of the ship's completed voyage and not for the entire voyage completed by the ship, the file is preliminary/non-final. I have specifically indicated that this is a preliminary IMMA data file for this ship by including the designation of "Prelim" in the name of the file. The name of the file is Jamestown_1844_Prelim_June_1861_Through_September_1865_Only.imma, and I have uploaded this particular Jamestown_1844_Prelim_June_1861_Through_September_1865_Only.imma file to GitHub. This preliminary IMMA data file has been uploaded to GitHub, and it is now available and ready to be accepted/added into the OldWeather3 Master Repository from my LawrenceSpencer Master Repository on GitHub. Just to reiterate, this file is a preliminary/non-final IMMA data file for the OldWeather3 Jamestown (1844) ship because it contains the quality-controlled geographical positions and meteorological observations data for only the last 4 years and 3 months (for only the time period of June 1861 through September 1865) of the OldWeather3 Jamestown (1844) ship's completed voyage and not for the entire voyage completed by the ship.

Kindest Regards,
Larry Spencer